% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{okcolors}
\alias{okcolors}
\title{Retrieve a Color Palette from the okcolors package}
\usage{
okcolors(palette = "obsession", n = 5, direction = 1)
}
\arguments{
\item{palette}{Character. Name of the palette to use. One of `"it"`, `"stone"`, `"skyscrapers"`, `"obsession"`, or `"moment"`. Default is `"obsession"`.}

\item{n}{Integer. Number of colors to return. Default is `5`.}

\item{direction}{Integer. If `1` (default), returns the palette as is. If `-1`, reverses the order.}
}
\value{
A character vector of hex color codes of length `n`.
}
\description{
Returns a vector of `n` colors from a specified palette included in the okcolors package.
}
\examples{
okcolors("stone", 3)
okcolors("obsession", 7, direction = -1)

}
