% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_palettes}
\alias{valid_palettes}
\title{List Available Palettes in the okcolors Package}
\usage{
valid_palettes()
}
\value{
A character vector with the names of available palettes.
}
\description{
Returns a character vector of all valid palette names supported by the `okcolors` package.
}
\examples{
valid_palettes()
}
